[Refactor] refactoring for ATSP
